requirements.txt
Mezzanine==1.4.16
South==0.8.2
argparse==1.2.1
django-appconf==0.6
django-compressor==1.3
django-extensions==1.2.4
six==1.4.1
wsgiref==0.1.2
Django==1.5.4
Pillow==2.2.1
Werkzeug==0.9.4
bleach==1.2.2
django-bootstrap-form==3.0
filebrowser-safe==0.2.31
grappelli-safe==0.2.22
html5lib==0.95
oauthlib==0.6.0
psycopg2==2.5.1
pytz==2013.7
requests==1.2.3
requests-oauthlib==0.3.3
django-templated-email==0.4.9
pygeoip==0.2.7
